requirements.txt
Django>=5.0,<5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-environ>=0.11
Pillow>=10.3
cloudinary>=1.38
django-cloudinary-storage>=0.3.0
psycopg2-binary>=2.9
whitenoise>=6.6
gunicorn>=22.0
